refactor: deleting unnecessary data and refactoring
