--- public/data/homepage.json
@@ -1,30 +1,0 @@
-{
-  "data": {
-    "id": 1,
-    "attributes": {
-      "title": null,
-      "description": null,
-      "createdAt": "2024-07-20T22:25:25.687Z",
-      "updatedAt": "2024-07-21T12:07:23.006Z",
-      "publishedAt": "2024-07-20T22:26:45.508Z",
-      "locale": "en",
-      "hero": {
-        "id": 1,
-        "title": "Lorem ipsum dolor sit amet,consectetur consequat.",
-        "description": "Lorem ipsum dolor sit amet,consectetur consequat,lorem ipsum dolor sit amet,consectetur consequat.",
-        "backgroundImage": {
-          "data": {
-            "id": 12,
-            "attributes": {
-              "name": "car.png",
-              "alternativeText": null,
-              "url": "/uploads/safaricar.png",
-              "width": 1704,
-              "height": 1080
-            }
-          }
-        }
-      }
-    }
-  }
-}
