+++ src/app/page.tsx
@@ -1,15 +1,16 @@
-import homepage from "./../../public/data/homepage.json";
-import HeroContent from "./components/HeroContent/Index";
-import { Hero } from "@/types/Hero";
+import heroData from "./../data/heroData.json";
+import HeroContent from "./components/HeroContent/index";
 
 export default function Home() {
-  const hero = homepage.data.attributes.hero;
   return (
     <div>
       <div className="flex items-center justify-center">
         <h1 className="text-center font-bold text-4xl">Experience Rwanda</h1>
       </div>
-      <HeroContent {...(hero as Hero)} />
+      <HeroContent
+        imageUrl={heroData.imageUrl}
+        content={{ title: heroData.title, description: heroData.description }}
+      />
     </div>
   );
 }
