+++ src/data/heroData.json
@@ -1,0 +1,6 @@
+{
+  "id": 1,
+  "imageUrl": "/uploads/safaricar.png",
+  "title": "Welcome to ExperienceRw",
+  "description": "Welcome to ExperienceRwanda, your trusted travel partner. We offer scheduled trips to popular destinations in Rwanda, ensuring a seamless and enjoyable journey every time."
+}
